refactor: Move records search-related endpoints to a separate module
